[cipd] Add field to DescribeInstance for metadata
This change adds a `DescribeMetadata` field to the
`DescribeInstanceRequest` message. This is to be used to request the
metadata of an instance. The returned value will be the same as
`ListMetadata`, but it allows for one roundtrip instead of two. This
will be used by the `describe` function in the `cipd` CLI client.

Bug: https://crbug.com/1506737

Change-Id: Ic5ecd94ccb92932318b2a6e06672871fb23fc1d4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5077232
Auto-Submit: Chris Lewis <cflewis@google.com>
Commit-Queue: Chris Lewis <cflewis@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>