[milo] Adjust test name truncation behaviour.
Example: https://ci.chromium.org/ui/p/infra/builders/ci/luci-go-continuous-bionic-64/7805/test-results

The current truncation leads to tests being truncated at
different points, which makes it harder to scan a list of
test results. For long test names, this is unavoidable.

But on some builds, most test names are not that long, and we
are truncating them to widths that are 20% of the screen width.
Give them an opportunity to display in their full length.

BUG=None
TEST=None

Change-Id: If155b285442e472c40df59aad48dd63b2fb42226
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5075933
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>