[swarming] Partially parse and validate pools.cfg.
The goal for now is to implement ACLs, and for that we
need `Realm` field in pools.cfg which is relatively easy
to get.

A more complicated task of validating and "expanding" task
templates is left as a TODO. We'll need it when we start
scheduling tasks through Go.

R=jonahhooper@google.com

Change-Id: I11985d65873826101157e0f856432f200c8c695c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5090898
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Jonah Hooper <jonahhooper@google.com>