cv: stop explicit support quick dry run
The additional mode is defined in a generic way in the Project config.
However, LUCI CV has the assumption that the additional mode would just
be QUICK_DRY_RUN and convert it to an enum inside the LUCI CV code.
This change removes the restriction to allow arbitrary run mode in
LUCI CV.

There is one noticible behavioral changes is that the additonal run
mode will no longer be exported to BQ and I think we can address it
when we design LUCI CV run based BQ table to deprecate the legacy one.

R=ddoman

Bug: 1483511
Change-Id: I0ab5df1076aeeb07f6ecfab3d7fca710609ceac6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5075014
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>