[ftt] Land just the directory-checking part of the test-ftt script
How to run the script:
$ ./test-ftt -i ~/cr -p 10

It should fail with "not yet implemented" on your machine.
If it didn't, then test-ftt is not flexible enough to handle the quirks
of different dev environments and needs to be changed.

This change is extracted from https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4997550.

See discussion in b:308999386 for more discussion about the portability of the original script.

BUG=b:304649677, b:308999386
TEST=make test

Change-Id: I847160ac133c48c29d6a7201d56d0eb41761ae73
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5010080
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Allen Li <ayatane@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>