cv: propose new tryjob data model in the rpc response
Right now, all tryjobs (the retries and the final attempt) are flattened
in the rpc response. The client would need to take extra step to figure
out which tryjob is the intermediate retry and which one is the final
attempt that contributes to the failure of the run. This change propose
to group the tryjob so that all attempts for the same tryjob/builder
will be contained in the same tryjob message.

R=ddoman, meiring

Change-Id: I606edd8880ff600281cfa0b718e02a94224332ac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4998267
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>