[GCEP] Create a new queue to drain VMs
Bug: b/306188743

Move functionality from manage-bot to new queues to relive the traffic.
This changelist creates a separate drainVM task and queue. This doesn't
replace the drainVM operation in manageBot. That will be done once
the functionality is verified. The task is only run once in 15 minutes
currently

Change-Id: Iac5d420451cbac4712cefcf259c6b9baa3de2b62
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5096268
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Deniz Kara <denizkara@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>