[buildbucket][resultdb] increase build timeout to 5 days
The original ask was for a 3 day timeout, but that is for the execution of the build. 
Since the scheduling timeout is 30 hours, the total build timeout request is 102 hours.
To keep things "easy" and consistent, I am setting the max to 5 days.

buildbucket v1's timeout was increased to the same to keep things consistent: https://crrev.com/c/5159167

Bug:1511096
Change-Id: I82df0dca858ad6d01b5d6066372bd339a9728a54
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5159107
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Randy Maldonado <randymaldonado@google.com>