[swarming] [bq] Create bq writer boilderplate
This cl creates a lot of boilerplate and tests. The idea is to mock many
clients before writing the actual exportTask.

The main decision made in this CL is that we do not need to wait until
the managedwriter.ManagedStream finishes writing rows with the
AppendRows call. As soon as we have the protos, we can continue
importing from datastore and wait for them to finish writing before we
finalize the writeStream.

Hence the bqWriter object which will be passed as a parameter to the
actual exportTask.

Bug: 1500695
Change-Id: I94c8c465ba38ad6f9cbdcc01c42870bf03dbd280
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5104766
Reviewed-by: Sergey Fetisov <sfetisov@google.com>
Commit-Queue: Jonah Hooper <jonahhooper@google.com>