ui: Don't elide step summary text when expanded
When expand the step show all of the step summary in the expansion,
rather than only the second "line" onwards.

Also exposed a problem with the step summary display, in that the second
time it is expanded it will be empty as lit has moved the HTML nodes
from the computed observable.  We probably never noticed this because a
summary longer than one line is very rare.

Bug: chromium:1507073
Change-Id: I64febd4c504fc15c60db88b5bdde676f2901ae97
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5162112
Commit-Queue: Matthew Warton <mwarton@google.com>
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Auto-Submit: Matthew Warton <mwarton@google.com>