ui: Initial monitoring UI in LUCI UI based on SOM
This is in the labs section reflecting its incomplete status.  Currently
all information is coming from SOM, but over time it is intended to
remove the SOM backend completely.

Putting this in the UI now allows us to work on it iteratively and get
feedback from gardeners.

Bug: b:319315200
Change-Id: I869d6e7ebbaf395ea2db13525cff5f9d396b809c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5148842
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>