[datastore] Refactor and fix Query.Less(...).
In many code paths it was giving up on first passing "==" check
for a field, ignoring subsequent fields. This resulted
in e.g. below incorrect behavior (and many other variants of it):

  a = NewQuery("Foo").Eq("Prop", "val").Offset(100)
  b = NewQuery("Foo").Eq("Prop", "val").Offset(200)
  a.Less(b) == false // WRONG

It was also a big monolithic function with pretty deep nesting
which was obscuring the bug. I also need to extend it to support
`NewQuery(...).In(...)` and it would resulted in ever bigger and
scarier function if it is left as is. So split it up in smaller
per-field functions.

R=sfetisov@google.com, randymaldonado@google.com
CC=anushruth@chromium.org

Change-Id: I467d88a0f2f9c2d18fdbdb3cb3bf3e606e660511
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5224978
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Sergey Fetisov <sfetisov@google.com>