[provenance] Add buildType document for SLSA v1
This change adds a buildType document that details a SLSA v1
specification. This is required as `buildDefinition.buildType` needs to
point to a human-readable documentation of the specification used.
Fuchsia is generating attestations against CIPD packages at SLSA v1, and
so this document needs to exist somewhere.

As CIPD is owned by Chrome, and the attestations should be usable by any
projects that rely on CIPD, it makes sense for this to live here.

Change-Id: I6912c532052890cc0c5fb0c94a9e97212d0f27f3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5100512
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chris Lewis <cflewis@google.com>
Commit-Queue: Allen Li <ayatane@google.com>
Reviewed-by: Adam Wu <zhenyuwu@google.com>