analysis: attribute skipped tests to failures in any milestone
Ideally, the skipped test results should be attributed to failures
that caused them. i.e. failures in the same milestone if there are enough samples, or failures in the previous milestone if there are not
enough samples. But given that luci-analysis does not know whether
the testing filtering was activated from failures from the current
milestone or the previous milestone, it's hard to write a query that
can attribute correctly.

Instead, we attribute skipped test results to failures in the any
milestone. This is only an issue when
1. the test is failing in a previous milestone for a different root
cause, and
2. the failures in the other milestone did not trigger test filtering
themselves (therefore already marked as "triggered test filtering").
Hopefully, this should be very rare. And we can adjust the attribution
logic further if this become an issue.

R=meiring@google.com, mwarton@google.com

Bug: b/312590738
Change-Id: Ifbd6e99b8af787fe328d44bb240620f091dc2c7d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5208370
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>