luci-ui: migrate bisection to use the generated prpc bindings
Also change the way we share the impmenetation of components between
analysis and test analysis. Instead of accepting a type that is mostly
a subset of another type, and add special cases in the components,
convert type to a general type before they are passed down to the shared
components. This is required as the generated proto Analysis/
TestAnalysis types are less compatible with each other because enums
are now numbers instead of strings.

R=beining@google.com, nqmtuan@google.com

Bug: 1504937
Change-Id: I5f3316a0981e3f2088122124fd2172222ea3b66b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5246887
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>