[LogAnalysis] Set up table schema
I also renamed the old row to TextArtifactRowLegacy because I want to use the name TextArtifactRow for the new one.

Some of the vars are unused in this CL, but will be used in later CL, so I'll just ignore the Tricium complains and leave them there.

Bug:b/322254047
Change-Id: I31a7a4c6437a6587434e22834bf419ae35473094
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5256377
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>