[luciexe/build] use atomic.Int64
On 32-bit ports it's the caller's responsibility¹ to arrange for 64-bit
alignment of 64-bit words accessed atomically. State.buildPbVers wasn't,
leading to a panic on such systems. Use atomic.Int64, which is readable
and automatically aligned on 32-bit ports. (It was added in Go 1.19.)

¹ https://pkg.go.dev/sync/atomic#pkg-note-BUG

Change-Id: I8abec2e04fb17fd9466f3167581c3991ff189943
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5268803
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>