[LogAnalysis] Fix clustering
In BigQuery, clustering is limit to a maximum of 4 columns. We have 5 at the moment.

I created another column that combines the artifact_id and shard_id. Both fields are small, so the size of the column be insignificant. (I thought about removing the artifact_id and shard_id column, but decided to leave them for the convenience of querying).


Bug:b/322254047
Change-Id: If338c4d1da4207485fefa3c45d99330a8753f7fd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5299234
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>