[paniccatcher] Add a unit test to Catch
Earlier today (2024-02-15) I wanted to know whether Catch would run the
callback with a nil argument (like finally) or not run the callback at all when
it is being used inside a function body where no exception is thrown, such as
in the example below.

```
func () {
  defer Catch(...)
  fmt.Printf("%s\n", "hi") // do something that doesn't throw an exception.
}()
```

As it turns out, Catch does not run the callback at all when no exception is
thrown.  Let's add a dedicated test case for this to make it more obvious to
people digging through the code.

Change-Id: Ieb698db1a8e04e68999ca8475b2cbae6358db4f7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5299875
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>