[datastore] Optimize deduplication for key-ordered RunMulti queries.
If queries are ordered only by key, all duplicates will be
returned from the heap one after another and we can use a simple
check to skip them. This is important for CountMulti(...) that can
be visiting tens of thousands of entities: storing them all in
a hash map for deduplication is a waste of memory.

R=chanli@chromium.org, randymaldonado@google.com

Change-Id: I66d74c462760b76eb407e4fe93255090c6b123a7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5311750
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>