luci-ui: fix the current page being incorrectly unloaded in some cases.
Unmounting the <Outlet /> when navigating state is loading causes the
app to do unmount then remount the current page when search params are
updated.
This cause all the state in the current page to be discarded and hooks
to be rerun in the next cycle. This is bad for performance and
consistentcy reasons. And even worse, it can lead to infinite reload if
the search param is set when the page is mounted, leading to an
infitely mount -> update search param -> unmount -> mount cycle.

R=beining@google.com, mdraz@google.com

Bug: None
Change-Id: If11dd1c877c28188191e53b9eabf6a2f6a889b2f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5312254
Commit-Queue: Beining Chen <beining@google.com>
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Beining Chen <beining@google.com>