[cv] Include quota_exhaustion_msg field within UserLimit
This change adds quota_exhaustion_msg field to UserLimit message. This
allows admins to configure additional custom message that they could
append to the end of the quota quota message.

This field is added to UserLimit which means that the admins would have
configure it for each of the limit that they add even though the general
use-case might be to have a single custom message for a whole project.
This is done to just keep with the current design and since we can
always add wrappers within starlark code.

Bug: 41486138
Change-Id: I83b9b8f7da760a1245bf375da838d235d8d5fb11
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5314296
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Aravind Vasudevan <aravindvasudev@google.com>