[cv] Replace findRunPolicy with findRunLimit
This change replaces findRunPolicy() logic with findRunLimit(). With
supporting unlimited quota and changes in requirement such as supporting
custom pending message, the rest of the quota manager logic and the
caller logic needs access to the user_limit used by the run. This change
simply breaks findRunPolicy() to only identify the user_limit to use
which allows the caller to use config such as `unlimited` and
`quota_exhaustion_msg` directly.

A follow-up will update debit and credit interface to return the used
user_limit. This allows the caller to use it in cases such as appending
custom message to the pending message and its logs.

Bug: 41486138
Change-Id: Iddae241ebfd6bcfe8735cf729c0a1b5e70811d22
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5315964
Commit-Queue: Aravind Vasudevan <aravindvasudev@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>