[swarming] Emit task's result ASAP in `collect`.
Do not buffer them all before writing them to stdout or to
the JSON output. This causes OOMs. Note that it means the
order in which we output tasks is no longer preserved
(whatever finishes first ends up in the stdout and JSON
first).

For JSON output it should not be a problem at all, since
results are represented by a dict {taskID: <result>}, which
is already unordered. For stdout, we'll see if anyone notices...
There are prons and cons.

There's a caveat: when using `-task-summary-python` legacy flag,
we need to buffer JSON results, since the order of outputs is
important in that case (it must match the order of `taskIDs`).

R=chanli@chromium.org, randymaldonado@google.com
BUG=40276579

Change-Id: I25f6724ab46c8dbda904cdd908610e1390ad312d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5331989
Reviewed-by: Randy Maldonado <randymaldonado@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>