[swarming] Limit number of concurrent result fetches.
To avoid OOMing when fetching large stdouts. Together
with "streaming" of JSON results, this should reduce
peak RAM usage of existing `swarming collect` calls.

Note that we still fully buffer stdout of a single
task in memory before writing it to JSON (because we
can't write multiple JSON fields concurrently when
fetching multiple tasks: we need to fully prepare
a JSON record before we can write it).

This will be improved in a separate CL which will
introduce writing task's stdout to a file.

Also cleanup how context cancellation errors are logged
and reported in the JSON output.

R=chanli@chromium.org, randymaldonado@google.com
BUG=40276579

Change-Id: I644c3bd5c169361a5e9e762f9f2093991d049e2b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5333999
Reviewed-by: Randy Maldonado <randymaldonado@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>