cv: include structured tryjob failure reason in tryjob execution state
Before, if a run fails because of tryjob failure, Tryjob manager
will evaluate the message to be posted on the CL using template
language because the message would be different on each CL as the
message contains a link to the checks tab of each CL. This has caused
issue in the past that the summary markdown may accidentally contains
the syntax of template language causing template evaluation fails.

I'd like to change the flow to Tryjob manage simply record the failed
tryjob id in the execution state. When Run manager reads the execution
state, it will construct the message to CL from the ids and then post
the message to Gerrit.

R=ddoman

Change-Id: I69fee2e388bc0a77a12ebf071eddbdf80aa792f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5334002
Commit-Queue: Scott Lee <ddoman@chromium.org>
Auto-Submit: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Scott Lee <ddoman@chromium.org>