[swarming] Fix waiting for results, broken in the previous commit.
Previous commit changed `resultsCh` to be non-buffered. It means
if nothing is reading from it, writes to it block.

swarming.GetMany (in "WaitAll" mode) blocks until all tasks are
finished. It means we don't read from `resultsCh` until all the
waiting is done. Effectively it means we wait for tasks to finish
first, and only then fetch their results.

Everything still technically works, but concurrency is limited.

To truly do fetching and waiting at the same time, just run
swarming.GetMany itself in a goroutine (an alternative is to
do reading from `resultCh` in a goroutine, but it is more
complicated).

R=chanli@chromium.org, randymaldonado@google.com
BUG=40276579

Change-Id: I0d02a9c38791697a6599e8ee87518c1a90c0d3f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5337043
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>