[Analysis] Remove "expensive queries" permission.
A review indicated that this permission was not meaningfully
contributing to cost control, and was probably just harming the
user experience for open source contributors.

If we need to reintroduce this, a cria group would likely be
the preferred means instead of a permission, as it is controlled
centrally and can be more rapidly updated.

BUG=None
TEST=Integration tests

Change-Id: I3c41fd3ca48821448cf18e403fcdc6fa25344cf4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5348160
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>