[auth] Don't overwrite scopes list when UseIDTokens is true.
This matters when running locally with UserCredentialsMethod
and with cached credentials for scopes [S1, S2, userinfo.email].

Previously, setting UseIDTokens to true was resetting the list of
necessary scopes to [userinfo.email], which resulted in getting
"login required" errors (because there are no cached credentials
for [userinfo.email]).

But it is perfectly fine to use [S1, S2, userinfo.email] for
generating ID tokens. All we need is userinfo.email to be in
this set. This avoid confusing "login required" error.

R=justinsuen@google.com

Change-Id: If5f51925510111e0d0a668f0c80a58128f504555
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5354710
Reviewed-by: Justin Suen <justinsuen@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>