[swarming] Implement most of ListBotTasks RPC.
It accepts a lot of combinations of filters and sort orders,
but in practice most of them do not work due to missing indexes
or because they make no sense for this RPC. Report all these
as InvalidArgument errors with, hopefully, useful messages.

Separately, the returned protos still lack a bunch of important
fields, since TaskToRun entity doesn't have them:
https://chromium-review.googlesource.com/c/5368660

Finally, IncludePerformanceStats is not implemented yet. It will be
done in a separate CL, since doing it in the most efficient way is
somewhat non-trivial.

R=chanli@chromium.org

Change-Id: I21fda7bfe7caa97f9bc2784019c03a813870b132
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5366244
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>