luci-auth: Add -json-format flag
This flag allows luci-auth to print the token in various JSON formats,
which are supported by tools such as reclient or Bazel.

With the `reclient` format, luci-auth can easily be used as a
credential helper for reclient, allowing Chromium developers to use
authenticate with RBE without needing to install the gcloud SDK.

The `bazel` format is defined in this design doc:
https://github.com/bazelbuild/proposals/blob/main/designs/2022-06-07-bazel-credential-helpers.md

It is is expected to be used by `reclient` in the future, as they want to standardize on this format for all tools in the remote build execution ecosystem as part of b/328288625 (make it easy to implement luci-based authentication).

Bug: b/328289019
Change-Id: Iaa6c1a78c440b8c29e3a8ccac4509b761e5e666f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5369304
Auto-Submit: Philipp Wollermann <philwo@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Philipp Wollermann <philwo@chromium.org>