[LogAnalysis] Add Test Status to Artifact
The current approach [1] does not work for most (80%) of artifacts in staging because when BatchCreateArtifacts endpoint is called, the containing TestResult is not guaranteed to be created, so we cannot get the status from Test Result. This is because TestResult and artifacts are uploaded using different channels.

I am trying a different approach:
1. In result sink ReportTestResults endpoint [2], we allow artifacts to be uploaded with test result status. This can be done because ReportTestResults already had the test status.
2. In BatchCreateArtifact endpoint, we can just use that status, instead of querying Spanner.

This will require update to the result sink (rdb stream binary), but no update to test harnesses will be required.

Another alternative is to set up a job to run after an invocation finalizes to set the status of existing rows, but it requires updating BQ rows, and I think is more involved than the above approach.

This CL perform step (2) above, the update to result sink will be in another CL.

[1] https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5373865
[2] https://source.chromium.org/chromium/infra/infra/+/main:go/src/go.chromium.org/luci/resultdb/sink/sink_server.go;l=183

Bug:b/322254047
Change-Id: I20e717033518247fb866c103b0b0a01052e21326
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5379674
Reviewed-by: Mostafa Draz <mdraz@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>