[swarming] Introduce `exporter` module.
We are going to use BigQuery Write API for BQ uploads.
This API has a limit on 1000 concurrent *connections* [1]. It
is not exactly clear what a connection is, but if it is a gRPC
connection, we'll have at least one of those per *process* that
ever needs to upload stuff to BQ.

Based on numbers for existing Python code, once all rewrites
and migrations are done, `default-go` will likely run hundreds
of processes. We'll have a serious risk of hitting the BQ limit if
we distribute BQ upload tasks across `default-go` processes.

So instead we'll have `exporter` module whose sole purpose would
be to do BQ exports. It should have much fewer number of processes.

Also move all server executables to be under server/cmd/...
for consistency. This will require special consideration when
deploying this change.

[1]: https://cloud.google.com/bigquery/quotas#write-api-limits

R=chanli@chromium.org

Change-Id: I9b1de3007fe912383e8ab76433fe931774eb6fca
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5384730
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>