[cloudbots] Start Swarming bot after startup script
For CloudBots, we rely on starting swarming-start-bot.service with an EnvironmentFile. This file is created by the Compute Engine native startup script which is started by google-startup-scripts.service.

We noticed that sometimes the Swarming bot starts before the end of the startup script. This change forces the swarming bot to always start after the end of the startup script.

BUG=b:329347968

Change-Id: I4d5b35097ab9ef7ba976c658ec016c0970fb9323
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5398568
Commit-Queue: JB Tamas-Leloup <jtamasleloup@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Bryan Ni <bniche@google.com>