[swarming] Add a helper for doing datastore => BQ row conversion.
This will be used as a building block in the overall BQ export
pipeline. If knows how to fetch entities from some time range
and convert them into BQ row protos (using to-be-written
conversion functions).

I considered making it fancy with goroutine pools and parallel
fetches, but decided it's not worth the effort for now, since
the export is not latency sensitive. If it becomes an issue,
Fetcher is the place that will have to be updated.

R=chanli@chromium.org

Change-Id: I4da107ada6bb3f114391cef688762e625e905800
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5394609
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>