cv: introduce honor_gerrit_linked_accounts to cv config
Today, this feature is only relevant to chromium and the infra projects.
If we make it by default on for all projects, the extra call makes to
Gerrit is a waste because the response is most likely empty. Also, we
would need to make sure we've configured the Gerrit host to grant the
project scoped account access to view secondary email.

https://gerrit-review.googlesource.com/Documentation/access-control.html#capability_modifyAccount

Bug: 40287772
Change-Id: Ic9445f77289f695c90ab90a64b84f377949db761
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5444769
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>