[Buildbucket] UpdateBuildTask: backfill backend task if the task is forcefully canceled
In the flow that a user cancels a build -> BB waits up the grace_period+MinUpdateBuildInterval -> BB forcefully cancels the build and the backend task. The build is set to CANCELED with the backend task still in running state (because backend task cancelation is done asynchronously).

In this case BB should backfill the backend task to set its ended state when getting the UpdateBuildTask message - even if the build has been set to ended.

Change-Id: I9bbe139e5be4dde3a347a62c6baa5c59c1fbf39d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5449901
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Commit-Queue: Chan Li <chanli@chromium.org>