[auth-service] Cache snapshot of all AuthGroups
This CL adds caching of all AuthGroup entities. The revision of
the AuthDB when the AuthGroups were last fetched is recorded,
so that the latest AuthGroups are always returned, but not by
always hitting Datastore.

These changes should majorly speed up the responsiveness of the
UI (particularly the groups page, and the lookup page).

Bug: b:304805492
Change-Id: I2e2774264fab34f18c58b394d696887c7b5872b6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5449142
Commit-Queue: Anne Redulla <aredulla@google.com>
Reviewed-by: Rachael Newitt <renewitt@google.com>