logdog: put counters behind mutex
Within a single beam worker, the counter are not thread-safe.
This can cause the beam worker to restart.
Previously, we did not observe this issue because the RPC calls makes
it unlikely that updates to the counters will happen at the same time.
Now this can happen a fair bit when there's no entity to update in a
batch (when the key range is already processed in a previous attempt).

R=meiring@google.com

Bug: b/327302650
Change-Id: I31a9a3b58378a7db131921cb1bb6bac357ce93d2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5460822
Reviewed-by: Patrick Meiring <meiring@google.com>
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>