[cv] fix a bug in cltriggerer
It spawns a goroutine to check if the origin CL is still with the CQ
vote, and signals the child-vote workers to cancel its work if
the origin CL no longer holds the CQ vote.

However, the condition was wrong, so that
1. the goroutine wrongly signals the worker
2. the worker cancels the work
3. when PM receives the CL trigger task result back, it finds that
some(or all if unlucky) of the votes were cancelled, but finds that
the origin CL still holds the CQ vote.
4. therefore, PM schedules another cltriggerer task for the unvoted
child CLs.
5. In the next try, cltriggerer will try to vote as many child CLs
as possible in the 4 seconds.
6. eventually, cltriggerer would be able to vote all the child CLs
and PM will create runs for them.

Change-Id: I568dc29b95be1bc15462ed732b6598f716ee861c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5463299
Auto-Submit: Scott Lee <ddoman@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>