[auth-service] Tidy groups page
This CL removes the useless search button from the groups page. It is
unnecessary because group matching is triggered by input into the
search box. v1 has this button but there is no listener for it.

Also adds the link to a group's lookup page, and tooltips for those
links.

Bug: b:304805492
Change-Id: I56322da4e7ce7d243f6f830bef19e720a9918afe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5460824
Reviewed-by: Rachael Newitt <renewitt@google.com>
Commit-Queue: Anne Redulla <aredulla@google.com>