[swarming] Produce backward compatible JSON for BQ rows.
It will be needed for PubSub pushes done by the BQ exporter.

Unfortunately protobuf schema that BigQuery Write API wants
is not compatible with schema of PubSub messages. This was not
a problem for Python exporter because it was using deprecated
JSON-based BigQuery API and could use custom proto => JSON
converter. But in Go we must use the schema matching BQ rows.
In particular, JSON-valued fields are `string` and fields
representing durations are `double`.

Convert such fields back to structpb.Struct{} and
durationpb.Duration{} for PubSub export.

R=chanli@chromium.org

Change-Id: Idff389b88ff12b8add4a748bf44a5429ab101a9f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5492881
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>