[server] Pass scope set when minting OAuth token via GCE MD server.
This code path was originally written for the GKE environment, which
doesn't support requesting tokens with arbitrary set of scopes. At the
time, on GKE (just like on GCE) you could only get a token with some
predefined set of scopes (e.g. scopes set when the GCE VM is created).

Then this code was adapted as is for GAE and Cloud Run environments,
where it continued to request a token with the default set of scopes.

In most cases our servers just call Cloud APIs. The default token
we were getting has `cloud-platform` scope, which covers all cloud
APIs, so everything was working fine.

But there's a use case now for a scope not included in
`cloud-platform`. Before this CL there was no way to request
a token with such scope via LUCI APIs when running on GAE or
Cloud Run.

GAE and Cloud Run do not support configuring the predefined set
of scopes (like GCE or GKE does). Instead, the flavor of
the GCE metadata server used by GAE and Cloud Run accepts
a dynamic set of scopes via "/token?scopes=..." URI parameter.

This CL changes GAE and Cloud Run LUCI servers to always use the dynamic
mechanism of requesting scopes, since it seems to be working and it
is more flexible and supports custom scopes. It means our GAE and
Cloud Run servers may start using more narrowly scoped tokens
(and a bunch of them). In particular, they will support
non-cloud scopes.

The danger is that some existing code may be using some
bogus combination of scopes, which currently "works" because in
reality the default cloud-platform scope is used. Once this CL
lands, bogus scopes will be passed to the metadata server, which
may not like them.

This CL doesn't touch GCE or GKE environments, since the dynamic
mechanism definitely doesn't work on GCE, and I haven't researched
if it works on GKE now (the situation is complicated there, since
there are two metadata implementations on GKE: the default one and
the one used with GKE Workload Identities).

R=iannucci@chromium.org

Change-Id: I55928397c9dd40ad7e7d7382fbd7a51e71da3f12
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5499381
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>